bug fix for programmatic movement of slider
Previously when programmatically moving the sliders (when showing 2
pointers), the pointers would not accurately update if during the move
one of the pointers intersected with the other.  In v2.4.1 the sliders
update accurately moving left to right, but not from right to left.
This fix allows for programmatic movement in both directions.